refactor(openrefine): extract shared query builder OpenRefineQueries
